requirements-core.txt
# NOTE
# This file is a pinned broad CPU-oriented environment.
# It is NOT the current recommended minimal PDF setup.
# For the current recommendation, use:
#   pip install -r requirements-recommended-pdf.txt
#
# ============================================================
# requirements-core.txt
#
# Pinned dependencies for the lightweight stack:
#   local · mistral · mathpix · deepseekocr · markitdown · html_local ·
#   auto · office (pptx/xlsx) · docling
#
# Install:  pip install -r requirements-core.txt
#
# For the full GPU stack (PaddleOCR / MinerU / Marker / CUDA
# torch) use requirements.txt instead.
# ============================================================

# ── core / project infrastructure ───────────────────────────
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.7
colorama==0.4.6
colorlog==6.10.1
distro==1.9.0
eval_type_backport==0.3.0
filetype==1.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
lxml==6.0.2
markdown-it-py==4.0.0
marko==2.2.1
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==25.0
pluggy==1.6.0
pydantic==2.12.4
pydantic-settings==2.7.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.0.1
pytz==2025.2
regex==2025.11.3
requests==2.32.3
rich==13.7.1
rpds-py==0.29.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tabulate==0.9.0
tiktoken==0.7.0
typer==0.12.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0

# ── local engine (PDF · DOCX · images · text) ───────────────
# Reads PDF pages, Word documents, and drives pytesseract OCR.
beautifulsoup4==4.14.2
pillow==12.2.0
pypdf==6.10.0
pytesseract==0.3.13
python-docx==1.2.0
soupsieve==2.8

# ── office extra  (PPTX · XLSX)  [pip install ".[office]"] ──
# python-pptx reads PowerPoint; openpyxl reads Excel workbooks.
et_xmlfile==2.0.0
openpyxl==3.1.5
python-pptx==1.0.2

# ── html_local engine ───────────────────────────────────────
# beautifulsoup4 / lxml already listed above.
# For best-quality article extraction install trafilatura:
#   pip install ".[html]"   →  trafilatura>=1.6,<2

# ── mistral engine ──────────────────────────────────────────
# Requires MISTRAL_API_KEY in .env
mistralai==1.9.11

# ── mathpix engine ──────────────────────────────────────────
# Requires MATHPIX_APP_ID and MATHPIX_APP_KEY in .env.
# Uses `requests` from the base stack; no extra Python package is needed.

# ── deepseekocr engine (SiliconFlow API) ────────────────────
# Requires SILICONFLOW_API_KEY in .env; pypdfium2 renders PDFs
# to images for the vision model.
openai==1.109.1
pypdfium2==4.30.0

# ── markitdown engine  [pip install ".[markitdown]"] ────────
# Install markitdown[pdf] separately; no pinned transitive deps
# are listed here because they vary by markitdown version.

# ── docling engine  [pip install ".[docling]"] ───────────────
# Heavy stack: transformer models + document-understanding libs.
# CPU torch (torch==2.2.2) is sufficient; for GPU acceleration
# replace with the appropriate torch+cuXXX wheel from
# https://download.pytorch.org/whl/
accelerate==1.11.0
antlr4-python3-runtime==4.9.3
dill==0.4.0
docling==2.62.0
docling-core==2.51.1
docling-ibm-models==3.10.2
docling-parse==4.7.1
Faker==38.0.0
filelock==3.19.1
fsspec==2025.9.0
huggingface-hub==0.36.0
invoke==2.2.1
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
latex2mathml==3.78.1
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.18
networkx==3.5
numpy==2.2.6
omegaconf==2.3.0
pandas==2.3.3
polyfactory==3.0.0
psutil==7.1.3
pylatexenc==2.10
python-dateutil==2.9.0.post0
PyYAML==6.0.3
rapidocr==3.4.2
referencing==0.37.0
rtree==1.4.1
safetensors==0.6.2
scipy==1.16.3
semchunk==2.2.2
shapely==2.1.2
sympy==1.14.0
tokenizers==0.22.1
# CPU torch for the CPU-oriented docling stack. Use requirements.txt for MinerU 3.x.
torch==2.2.2
tqdm==4.67.1
transformers==4.57.3
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-java==0.23.5
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
xlsxwriter==3.2.9
